This hides the routes for make a date and my dates but they're not protected.
